pyproject.toml
[tool.poetry]
name = "challenge-bravo"
version = "0.1.0"
description = "Hurb back-end test"
authors = ["avlm <victor.luckwu@gmail.com>"]
packages = [
    { include = 'challenge_bravo' }
]

[tool.poetry.dependencies]
python = "^3.8"
peewee = "^3.14.4"
fastapi = "^0.65.1"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
python-decouple = "^3.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
ipython = "^7.23.1"
ipdb = "^0.13.8"
flake8 = "^3.9.2"
pytest = "^6.2.4"
pre-commit = "^2.13.0"
pytest-cov = "^2.12.0"
requests = "^2.25.1"
locust = "^1.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
